my_trading_modules_20250604_045025/requirements.txt
websockets>=11.0.3
ccxt>=4.2.25
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
requests>=2.31.0
flask>=2.3.0
tabulate>=0.9.0